suppression du temps de transition quand change de vitesse + correction warnings

+++ MiAmusing/Source/AmusingSceneComponent.cpp
@@ -103,7 +103,7 @@ void AmusingSceneComponent::renderOpenGL()
 	case TextScale:
 		break;
 	case ColourButtons:
-		for (int i = 0; i < buttonsColor.size(); ++i)
+		for (int i = 0; i < (int)buttonsColor.size(); ++i)
 		{
 			g.setColour(buttonsColor[i]);
 			g.fillRect(getWidth() - 100, 4 + i * (getHeight() - 8) / buttonsColor.size(), 100 - 4, (getHeight() - 8) / (int)buttonsColor.size());
